+++ changelog.md
@@ -5,6 +5,7 @@
 - Added trees
 - Added new blocks: log and leaf
 - Add different textures per block face (used for log and grass)
+- Fixed orientation of side textures
 
 ## 0.1.4
 - New save file format (v2.0) using run-length encoding